# zstd: asm version of decodeSync