… zstd/decoder_test.go
… zstd/seqdec.go
… zstd/seqdec_amd64.go
… zstd/seqdec_amd64.s
… zstd/seqdec_generic.go