@WojciechMula 2022-04-08
> The current code appears stable and we can proceed with this one when you have the time.

I'm fighting right now with avo, after rebasing to the master. I wish your PR https://github.com/mmcloughlin/avo/pull/105 got merged.