@WojciechMula 2022-04-10
> @WojciechMula Yeah, I squeezed out the last registers, and it can sometimes we quite hard to figure out where allocs are failing.
> 
> If you need to free one or more registers, you can use:
> 
> ```
> 		seqsBaseStash := AllocLocal(8)
> 		MOVQ(seqsBase, seqsBaseStash)
> 		seqsBase = nil
> 		//
> 		// Code that doesn't need seqsBase
> 		//
> 		// reload:
> 		seqsBase = GP64()
> 		MOVQ(seqsBaseStash, seqsBase)
> ```
> 
> This will allocate stack for the value.

Thank you. I spilled some values, but still, have one place where allocation fails. I will work on it on Monday. I hope it will finally compile, but I'm afraid performance improvement will drop when we put too many values on the stack.