@klauspost 2022-04-19
>  I hope it will finally compile, but I'm afraid performance improvement will drop when we put too many values on the stack.

Don't worry too much about that. Stack is pretty much always in L1 and reads/writes can be done async to most other code.

Futhermore, some archs (Zen 2 for instance) can do memory to virtual register mapping, meaning that writing to non-changing addresses are kept in virtual CPU registers. There is so much magic going on with modern super-scalar CPUs :)

If you would like assistance, I can take a stab at finishing up the code. For now I will review the code and test it out.