@klauspost 2022-04-21
I think for now we should ditch the output reallocate logic.

For blocks that have the frame-content-size set we can allocate the output what we need, and we would never need to extend the output. We must fail these if we exceed frame-content-size anyway.

For decodes *without* frame content size we allocate the [default](https://github.com/klauspost/compress/blob/000e3ec3185f1cc3d2363c6bff4af11c176aca2c/zstd/decoder.go#L362) and let it fall back once we have less than a full frame left on output.